Update main.js
Compled the result() function:

Created a new variable called newStory, and set its value to equal storyText. 

Created three new variables called xItem, yItem, and zItem, and made them equal to the result of calling randomValueFromArray() 

Replaced the three placeholders in the newStory string — :insertx:, :inserty:, and :insertz: — with the strings stored in xItem, yItem, and zItem. 

Inside the first if block, added another string replacement method call to replace the name 'Bob' found in the newStory string with the name variable. 

Inside the second if block, replaced 300 with a calculation that converts 300 pounds into stones. Concatenate ' stone' onto the end of the result of the overall Math.round() call.

Inside the line that defines the temperature variable, replace 94 with a calculation that converts 94 Fahrenheit into centigrade. Concatenated ' centigrade' onto the end of the result of the overall Math.round() call.

Added two more string replacement lines that replace '94 fahrenheit' with the contents of the temperature variable, and '300 pounds' with the contents of the weight variable.
Finally, in the second-to-last line of the function, made the textContent property of the story variable (which references the paragraph) equal to newStory.