Update main.js
Declared a const array listing the filenames of each image.

Declared a const object listing the alternative text for each image.
Looping through the array of filenames, and for each one, inserted an <img> element inside the thumb-bar <div> that embeds that image in the page along with its alternative text.

Added a click event listener to each <img> inside the thumb-bar <div> so that, when they are clicked, the corresponding image and alternative text are displayed in the displayed-img <img> element.

Added a click event listener to the <button> so that when it is clicked, a darken effect is applied to the full-size image. When it is clicked again, the darken effect is removed again.