Update index.html
For The audio player,it  isn't accessible to hearing impaired (deaf) people, i added a mp4 file and removed the mp3 file. The mp4 file includes subtiles that follow what they're saying.

Because the <audio> player isn't accessible to those using older browsers that don't support HTML audio. I added a transcript at the bottom and a way for users to download the video